requirements.txt
pandas>=1.3.0
numpy>=1.21.0
faiss-cpu>=1.7.0
sentence-transformers>=2.0.0
scikit-learn>=1.0.0
torch>=1.9.0
transformers>=4.0.0
streamlit>=1.28.0
